+++ src/components/Home/Featured-Gallery/gallery-data.js
@@ -1,0 +1,19 @@
+import Image1 from "../../../assets/images/FB_IMG_1609818291318_Original.jpeg";
+import Image2 from "../../../assets/images/FB_IMG_1609818276266_Original.jpeg";
+import Image3 from "../../../assets/images/FB_IMG_1609818328585_Original.jpeg";
+import Image4 from "../../../assets/images/FB_IMG_1606604883160_Original.jpeg";
+import Image5 from "../../../assets/images/FB_IMG_1609818199940_Original.jpeg";
+import Image6 from "../../../assets/images/FB_IMG_1609818315477_Original.jpeg";
+import Image7 from "../../../assets/images/20210707_155857_Original.jpeg";
+import Image8 from "../../../assets/images/FB_IMG_1606604670891_Original.jpeg";
+
+export const images = [
+  { src: Image1, alt: "Exterior view of a modern house with large windows" },
+  { src: Image2, alt: "Newly renovated kitchen with marble countertops" },
+  { src: Image3, alt: "Spacious living room with hardwood floors" },
+  { src: Image4, alt: "Backyard patio area with outdoor furniture" },
+  { src: Image5, alt: "Master bedroom with en-suite bathroom" },
+  { src: Image6, alt: "Home office setup with built-in shelving" },
+  { src: Image7, alt: "Landscaped front yard with walkway" },
+  { src: Image8, alt: "Custom built-in storage solutions in garage" },
+];
