+++ src/components/Services-Gallery/services-list.js
@@ -1,5 +1,3 @@
-// services.js
-
 export const services = [
   {
     title: "Renovations",
