+++ src/css/details-page.css
