+++ src/css/hero-section.css
@@ -8,7 +8,7 @@
 .hero {
     text-align: center;
     padding: 4rem 2rem;
-    background: url('../../assets/images/FB_IMG_1609818258950_Original.jpeg') center/cover no-repeat;
+    background: url('../assets/images/FB_IMG_1609818258950_Original.jpeg') center/cover no-repeat;
     color: var(--text-color);
 }
