+++ src/css/services-gallery.css
@@ -158,7 +158,7 @@
 
 .image-grid {
     display: grid;
-    grid-template-columns: repeat(auto-fit, minmax(200px, 1fr));
+    grid-template-columns: repeat(auto-fit, minmax(75px, 1fr));
     gap: 20px;
     padding: 20px;
     width: 100%;
@@ -259,12 +259,14 @@
 
 @media screen and (max-width: 480px) {
     .navigation-btn {
-        width: 35px;
-        height: 35px;
+        display: none;
     }
 
     .navigation-btn svg {
-        width: 16px;
-        height: 16px;
+       display: none;
+    }
+
+    .image-grid {
+        grid-template-columns: repeat(auto-fit, minmax(150px, 1fr));
     }
 }
