+++ src/pages/Home.js
@@ -1,8 +1,8 @@
-import * as React from "react";
-import Hero from "../components/Home-Components/hero";
-import FeaturedServices from "../components/Home-Components/featuredServices";
-import FeaturedGallery from "../components/Home-Components/featuredGallery";
-import Testimonials from "../components/Home-Components/testimonials";
+import React from "react";
+import Hero from "../components/Home/hero-section";
+import FeaturedServices from "../components/Home/featured-services";
+import FeaturedGallery from "../components/Home/Featured-Gallery/featured-gallery";
+import Testimonials from "../components/Home/testimonials";
 
 const Home = () => {
   return (
